+++ core/src/apu/mod.rs
@@ -9,6 +9,7 @@ use self::{mixer::Mixer, noise::Noise, tone::Tone, wave::Wave};
 mod clock_divider;
 mod mixer;
 mod noise;
+mod sweep;
 mod timer;
 mod tone;
 mod util;
